refactor: move aliases to other modules
