--- yumako_jams/src/aliases.rs
@@ -1,9 +1,0 @@
-use futures::stream::BoxStream;
-
-use crate::radio_item::RadioItem;
-
-pub type LayerResult<'a> = Result<RadioStream<'a>, crate::Error>;
-
-pub type RadioResult = Result<RadioItem, crate::Error>;
-
-pub type RadioStream<'a> = BoxStream<'a, RadioResult>;
