+++ yumako_jams/src/json/layer.rs
@@ -4,13 +4,12 @@ use serde::Deserialize;
 use serde::Serialize;
 use serde_json::Value;
 
-use crate::aliases::LayerResult;
-use crate::aliases::RadioStream;
 use crate::client::YumakoClient;
 use crate::modules::filters::booleans::AndFilter;
 use crate::modules::filters::cooldown::CooldownFilter;
 use crate::modules::filters::minimum_listens::MinimumListenFilter;
 use crate::modules::filters::timeout::TimeoutFilter;
+use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
 use crate::modules::scores::listenrate::ListenRateScorer;
 use crate::modules::scores::overdue_count::OverdueCountScorer;
@@ -18,6 +17,7 @@ use crate::modules::scores::overdue_duration::OverdueDurationScorer;
 use crate::modules::scores::sort::SortModule;
 use crate::modules::seeders::listen_seeder::ListenSeeder;
 use crate::radio_variables::RadioVariables;
+use crate::RadioStream;
 
 /// A layer represent a step in the radio processing. It calls a module based on the step type
 #[derive(Serialize, Deserialize, Clone)]
