+++ yumako_jams/src/modules/filters/minimum_listens.rs
@@ -6,10 +6,10 @@ use futures::TryStreamExt;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::aliases::LayerResult;
-use crate::aliases::RadioStream;
 use crate::client::YumakoClient;
+use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
+use crate::RadioStream;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct MinimumListenFilter {
