+++ yumako_jams/src/modules/filters/timeout.rs
@@ -8,10 +8,10 @@ use futures::TryStreamExt;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::aliases::LayerResult;
-use crate::aliases::RadioStream;
 use crate::client::YumakoClient;
+use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
+use crate::RadioStream;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct TimeoutFilter {
