+++ yumako_jams/src/modules/scores/overdue_duration.rs
@@ -2,12 +2,12 @@ use alistral_core::datastructures::listen_collection::traits::ListenCollectionRe
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::aliases::LayerResult;
-use crate::aliases::RadioStream;
 use crate::client::YumakoClient;
+use crate::modules::radio_module::LayerResult;
 use crate::modules::radio_module::RadioModule;
 use crate::modules::scores::ScoreMerging;
 use crate::radio_stream::RadioStreamaExt as _;
+use crate::RadioStream;
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct OverdueDurationScorer {
